Former-commit-id: 2d08a056cf7c03fecd0673f3923667d2182a6e2b

+++ face/src/main/java/com/simprints/face/data/moduleapi/face/responses/entities/FaceMatchResult.kt
@@ -5,4 +5,4 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class FaceMatchResult(override val guid: String,
-                           override val confidence: Int) : IFaceMatchResult
+                           override val confidence: Float) : IFaceMatchResult
